Add simple photo action API.
This will be a requirement, all operations need to be undoable. Written in such
a way that an action is specified without giving the photo to which it should
be applied. This way we can apply it to a photo in the main cache and then
reapply it to the actual photo.