Add an ImageSurfaceCache for surface handover.
The icon view had flashes every time a new line was scrolled into view.
This was due to image surfaces being recreated. Introducing a new cache,
that's specific to the view widget (and can be disposed along with it)
solves this problem.

All rendering and cache updating is done on the main thread. This is
basically Alan McGovern's "how many threads do you need? zero!" trick.
And it works wonderfully well.