Implement a caching model based on enumerations
This commit adds a caching model based on enumerations. It can either
take a simple IEnumerable, which it will then walk and show (the dumb
approach) or it can accept a TripodQuery, which is an IEnumerable that
can translate into the relevant bits needed for ICacheableDatabaseModel.

This has two approaches: we can handle each IEnumerable, regardless of
where it's from and we can do highly optimized SQL caching for database
backed sources.

There's still a couple of caveats: it looks as if the extension methods
of IEnumerable get a higher priority than the implementation methods in
TripodQuery. Need to investigate.